+++ RELEASENOTES.md
@@ -1,3 +1,11 @@
+## [0.2.0]
+
+### Added
+- add control panel
+- add editing
+- add write to server method
+- add event click
+
 ## [0.1.0]
 
 ### Added